requirements.txt
APScheduler==3.6.3
argcomplete==1.12.2
asgiref==3.3.1
boto==2.49.0
boto3==1.16.44
botocore==1.19.44
certifi==2020.6.20
cffi==1.14.4
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-axes==5.10.0
django-cors-headers==3.6.0
django-email-verification==0.0.7
django-extensions==3.1.0
django-google-maps==0.12.2
django-ipware==3.0.2
django-reset-migrations==0.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-multiple-models==2.1.3
django-seed==0.2.2
django-storages==1.6.6
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-api-tracking==1.7.8
durationpy==0.5
Faker==5.1.0
future==0.18.2
googlemaps==4.4.2
hjson==3.0.2
idna==2.10
jmespath==0.10.0
kappa==0.6.0
numpy==1.19.4
oauthlib==3.1.0
pandas==1.2.0
Pillow==8.0.1
pip-tools==5.4.0
placebo==0.9.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.55.0
troposphere==2.6.3
tzlocal==2.1
urllib3==1.25.10
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0